+++ examples/benchmarks.py
@@ -59,9 +59,9 @@ db = Database('perfDB')
 table = db.table('sessions')
 table.index('by_multiple', '!{origin}|{day:02}|{hour:02}|{sid:05}')
 chunk(table, 0, 5000)
-print("")
-for doc in table.find('by_multiple', limit=10):
-    print('{origin} {day} {hour} {sid} {when}'.format(**doc))
+#print("")
+#for doc in table.find('by_multiple', limit=10):
+#    print('{origin} {day} {hour} {sid} {when}'.format(**doc))
 
 
 start = 0
@@ -69,7 +69,7 @@ count = 5000
 begin = time()
 for doc in table.find('by_multiple'): pass
 finish = time()
-print("  - {:5}:{:5} - Append Speed/sec = {:.0f}".format(start, count, count / (finish - begin)))
+print("  - {:5}:{:5} - Read Speed/sec = {:.0f}".format(start, count, count / (finish - begin)))
 
 db.close()